Camera Protoype code should be working (unable to verify until dev account set up)
The new view is of Class type CameraViewController.

It has 3 methods that pertain to each button and others that pertain to
what is called the Image Picker that gets the image from the Camera UI
and applies it to the UIImageView in the center of the prototype. This
will probably change as I believe the camera UI should launch to begin
with and then everything handle from there but I dont know how possible
this is or isnt.