Added preview images for filters that have been created see description for more info
The PSD is available here: http://d.pr/f/ida1

Once a filter is made, open the iOS Simulator and in Safari go to this
link and download the full size image to the Camera Roll:
http://d.pr/i/z4EZ

Open the HARN app and navigate to the new filter and choose the above
image as the one you wish to apply a filter for. Save and send it
yourself to place inside of the PSD. Export that png and then add it to
the Project.

In filterview.m there is a list of areas where Preview Images are made.
Change the case condition to the filter's name and apply the image as
the ImageView image there.