Resolved a lot of auto-layout issues in regard to detail view controller and title view controller. See description for more info
Title View continues to be a work in progress:
Bugs:
1) animations dont translate properly in scrollview
2) when returning to the view the list of collections is not shown
anymore and the gallery is improperly pushed over it.
3) Titles of the gallery works is hard to read on the wood for 3.5"
displays, and the page dots do not show on 3.5" displays.

Info button in Detail view matches the tint color of the navigation bar
items. The text and images in detail view now autoresizes to conform to
both 3.5 and 4" screen sizes. ALL XCODE WARNINGS THAT ARE CODE RELATED
ARE RESOLVED.

BUGS:
1) when returning from another view (share screen or info view) and the
title had more than one line, it will be cut off.
2) The tint Color of the info button does not get sent over to the
information view's bar buttons  properly
3) resize of image happens jarringly (not a major priority but
something to think about)