More bug fixes for front page
Fixed an issue that might cause the page control dots on the bottom of
the page to shift after more info per gallery has been expanded.

iOS devices running 3.5" screens will have just a white background on
the gallery page. it looks nice and resolves a lot of visual problems
related to resizing the image and color of text.